Store iterations count in Long to reduce possibility of an overflow
Closes #179

+++ runtime/nativeMain/src/kotlinx/benchmark/native/NativeExecutor.kt
@@ -65,7 +65,7 @@ internal class NativeExecutor(
         val (configFileName, iteration, cycles, resultsFile) = additionalArguments
         val benchmarkRun = configFileName.parseBenchmarkConfig()
         val benchmark = benchmarks.getBenchmark(benchmarkRun.benchmarkName)
-        val samples = run(benchmark, benchmarkRun, iteration.toInt(), cycles.toInt())
+        val samples = run(benchmark, benchmarkRun, iteration.toInt(), cycles.toLong())
         resultsFile.writeFile(samples?.let{ it[0].toString() } ?: "null")
     }
 
@@ -132,7 +132,7 @@ internal class NativeExecutor(
         benchmark: BenchmarkDescriptor<Any?>,
         benchmarkRun: BenchmarkRun,
         currentIteration: Int? = null,
-        cyclesPerIteration: Int? = null
+        cyclesPerIteration: Long? = null
     ): DoubleArray? {
 
         require((currentIteration == null) == (cyclesPerIteration == null)) {
@@ -253,7 +253,7 @@ internal class NativeExecutor(
     }
 
     private inline fun measureSingleIterationLoop(
-        cycles: Int,
+        cycles: Long,
         nativeGCAfterIteration: Boolean,
         body: () -> Unit,
     ): Double {
@@ -279,7 +279,7 @@ internal class NativeExecutor(
         benchmark: BenchmarkDescriptor<T>,
         workers: WorkersPool,
         currentIteration: Int? = null
-    ): Int {
+    ): Long {
         require(
             config.nativeFork == NativeFork.PerIteration && currentIteration != null
                     || config.nativeFork == NativeFork.PerBenchmark && currentIteration == null
@@ -288,7 +288,7 @@ internal class NativeExecutor(
         }
 
         val warmupIterations = if (config.nativeFork == NativeFork.PerIteration) 1 else config.warmups
-        var iterations = 0
+        var iterations = 0L
         repeat(warmupIterations) { iteration ->
             val benchmarkNanos = config.iterationTime * config.iterationTimeUnit.toMultiplier()
 
@@ -307,12 +307,12 @@ internal class NativeExecutor(
         iterationDurationNanos: Long,
         nativeGCAfterIteration: Boolean,
         body: () -> Unit,
-    ): Pair<Int, Double> {
+    ): Pair<Long, Double> {
         if (nativeGCAfterIteration)
             GC.collect()
         val startTime = TimeSource.Monotonic.markNow()
         var duration = Duration.ZERO
-        var iterations = 0
+        var iterations = 0L
         while (duration.inWholeNanoseconds < iterationDurationNanos) {
             body()
             duration = startTime.elapsedNow()
@@ -330,7 +330,7 @@ internal class NativeExecutor(
         nativeGCAfterIteration: Boolean,
         iterationDurationNanos: Long,
         workers: WorkersPool,
-    ): Pair<Int, Double> {
+    ): Pair<Long, Double> {
         return singleIteration(instance, benchmark, workers, { results ->
             results.sumOf { it.first } / results.size to results.sumOf { it.second }
         }) { body ->
@@ -343,7 +343,7 @@ internal class NativeExecutor(
     private fun <T> measureSingleIteration(
         instance: T,
         benchmark: BenchmarkDescriptor<T>,
-        cycles: Int,
+        cycles: Long,
         nativeGCAfterIteration: Boolean,
         workers: WorkersPool,
     ): Double {
